+++ test/test-change-both.sh
@@ -261,6 +261,19 @@ if [ -z "$match" ]; then
     exit 1
 fi
 
+# Cannot pull if edits in candidate
+new "edit local candidate"
+expectpart "$($clixon_cli -1f $CFG -m configure delete devices device openconfig1 config interfaces interface z)" 0 "^$"
+
+new "pull expect error"
+expectpart "$($clixon_cli -1f $CFG pull replace 2>&1)" 255 "Cannot pull if the candidate datastore is modified"
+
+new "discard"
+expectpart "$($clixon_cli -1f $CFG -m configure discard)" 0 "^$"
+
+new "pull again"
+expectpart "$($clixon_cli -1f $CFG pull replace 2>&1)" 0 "OK"
+
 if $BE; then
     new "Kill old backend"
     stop_backend -f $CFG